Refactor the new utils package
