+++ cmd/lifecycle/detector.go
@@ -7,10 +7,10 @@ import (
 	"github.com/buildpacks/lifecycle"
 	"github.com/buildpacks/lifecycle/buildpack"
 	"github.com/buildpacks/lifecycle/cmd"
+	"github.com/buildpacks/lifecycle/encoding"
 	"github.com/buildpacks/lifecycle/platform"
 	"github.com/buildpacks/lifecycle/platform/common"
 	"github.com/buildpacks/lifecycle/priv"
-	"github.com/buildpacks/lifecycle/utils"
 )
 
 type detectCmd struct {
@@ -143,11 +143,11 @@ func (da detectArgs) verifyBuildpackApis(order buildpack.Order) error {
 }
 
 func (d *detectCmd) writeData(group buildpack.Group, plan platform.BuildPlan) error {
-	if err := utils.WriteTOML(d.groupPath, group); err != nil {
+	if err := encoding.WriteTOML(d.groupPath, group); err != nil {
 		return cmd.FailErr(err, "write buildpack group")
 	}
 
-	if err := utils.WriteTOML(d.planPath, plan); err != nil {
+	if err := encoding.WriteTOML(d.planPath, plan); err != nil {
 		return cmd.FailErr(err, "write detect plan")
 	}
 	return nil
