+++ imageutils/imageutils_test.go
@@ -1,4 +1,4 @@
-package utils_test
+package imageutils_test
 
 import (
 	"io/ioutil"
@@ -11,8 +11,8 @@ import (
 	"github.com/sclevine/spec/report"
 
 	"github.com/buildpacks/lifecycle/buildpack"
+	"github.com/buildpacks/lifecycle/encoding"
 	h "github.com/buildpacks/lifecycle/testhelpers"
-	"github.com/buildpacks/lifecycle/utils"
 )
 
 func TestUtils(t *testing.T) {
@@ -20,31 +20,6 @@ func TestUtils(t *testing.T) {
 }
 
 func testUtils(t *testing.T, when spec.G, it spec.S) {
-	when(".TruncateSha", func() {
-		it("should truncate the sha", func() {
-			actual := utils.TruncateSha("ed649d0a36b218c476b64d61f85027477ef5742045799f45c8c353562279065a")
-			if s := cmp.Diff(actual, "ed649d0a36b2"); s != "" {
-				t.Fatalf("Unexpected sha:\n%s\n", s)
-			}
-		})
-
-		it("should not truncate the sha with it's short", func() {
-			sha := "not-a-sha"
-			actual := utils.TruncateSha(sha)
-			if s := cmp.Diff(actual, sha); s != "" {
-				t.Fatalf("Unexpected sha:\n%s\n", s)
-			}
-		})
-
-		it("should remove the prefix", func() {
-			sha := "sha256:ed649d0a36b218c476b64d61f85027477ef5742045799f45c8c353562279065a"
-			actual := utils.TruncateSha(sha)
-			if s := cmp.Diff(actual, "ed649d0a36b2"); s != "" {
-				t.Fatalf("Unexpected sha:\n%s\n", s)
-			}
-		})
-	})
-
 	when(".WriteTOML", func() {
 		var tmpDir string
 
@@ -62,7 +37,7 @@ func testUtils(t *testing.T, when spec.G, it spec.S) {
 
 		it("should write TOML", func() {
 			group := buildpack.Group{Group: []buildpack.GroupBuildpack{{ID: "A", Version: "v1"}}}
-			if err := utils.WriteTOML(filepath.Join(tmpDir, "subdir", "group.toml"), group); err != nil {
+			if err := encoding.WriteTOML(filepath.Join(tmpDir, "subdir", "group.toml"), group); err != nil {
 				t.Fatal(err)
 			}
 			b := h.Rdfile(t, filepath.Join(tmpDir, "subdir", "group.toml"))
