+++ rebaser.go
@@ -10,8 +10,9 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/buildpacks/lifecycle/api"
+	"github.com/buildpacks/lifecycle/imageutils"
 	"github.com/buildpacks/lifecycle/platform"
-	"github.com/buildpacks/lifecycle/utils"
+	"github.com/buildpacks/lifecycle/str"
 )
 
 type Rebaser struct {
@@ -25,7 +26,7 @@ type RebaseReport struct {
 
 func (r *Rebaser) Rebase(appImage imgutil.Image, newBaseImage imgutil.Image, additionalNames []string) (RebaseReport, error) {
 	var origMetadata platform.LayersMetadataCompat
-	if err := utils.DecodeLabel(appImage, platform.LayerMetadataLabel, &origMetadata); err != nil {
+	if err := imageutils.DecodeLabel(appImage, platform.LayerMetadataLabel, &origMetadata); err != nil {
 		return RebaseReport{}, errors.Wrap(err, "get image metadata")
 	}
 
@@ -80,7 +81,7 @@ func (r *Rebaser) Rebase(appImage imgutil.Image, newBaseImage imgutil.Image, add
 	}
 
 	hasPrefix := func(l string) bool { return strings.HasPrefix(l, "io.buildpacks.stack.") }
-	if err := utils.SyncLabels(newBaseImage, appImage, hasPrefix); err != nil {
+	if err := imageutils.SyncLabels(newBaseImage, appImage, hasPrefix); err != nil {
 		return RebaseReport{}, errors.Wrap(err, "set stack labels")
 	}
 
@@ -101,18 +102,18 @@ func validateMixins(appImg, newBaseImg imgutil.Image) error {
 	var appImageMixins []string
 	var newBaseImageMixins []string
 
-	if err := utils.DecodeLabel(appImg, platform.MixinsLabel, &appImageMixins); err != nil {
+	if err := imageutils.DecodeLabel(appImg, platform.MixinsLabel, &appImageMixins); err != nil {
 		return errors.Wrap(err, "get app image mixins")
 	}
 
-	if err := utils.DecodeLabel(newBaseImg, platform.MixinsLabel, &newBaseImageMixins); err != nil {
+	if err := imageutils.DecodeLabel(newBaseImg, platform.MixinsLabel, &newBaseImageMixins); err != nil {
 		return errors.Wrap(err, "get run image mixins")
 	}
 
-	appImageMixins = utils.RemoveStagePrefixes(appImageMixins)
-	newBaseImageMixins = utils.RemoveStagePrefixes(newBaseImageMixins)
+	appImageMixins = str.RemoveStagePrefixes(appImageMixins)
+	newBaseImageMixins = str.RemoveStagePrefixes(newBaseImageMixins)
 
-	_, missing, _ := utils.Compare(newBaseImageMixins, appImageMixins)
+	_, missing, _ := str.Compare(newBaseImageMixins, appImageMixins)
 
 	if len(missing) > 0 {
 		sort.Strings(missing)
