+++ save.go
@@ -9,7 +9,7 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/buildpacks/lifecycle/platform"
-	"github.com/buildpacks/lifecycle/utils"
+	"github.com/buildpacks/lifecycle/str"
 )
 
 func saveImage(image imgutil.Image, additionalNames []string, logger Logger) (platform.ImageReport, error) {
@@ -73,7 +73,7 @@ func (me *MultiError) Error() string {
 func shortID(identifier imgutil.Identifier) string {
 	switch v := identifier.(type) {
 	case local.IDIdentifier:
-		return utils.TruncateSha(v.String())
+		return str.TruncateSha(v.String())
 	case remote.DigestIdentifier:
 		return v.Digest.DigestStr()
 	default:
